yw-ww-backend/requirements.txt
# Minimum packages for azure functions
azure-functions
azure-functions-worker
azure-storage-blob==12.0.0
grpcio==1.14.1
grpcio-tools==1.14.1
protobuf==3.6.1
six==1.11.0

#additional analysis packages
numpy==1.18.2
pandas==1.0.3
networkx==2.4
xlrd == 1.2.0